@@ -1,3 +1,5 @@
+"use client";
+
 import { type ReactNode, useState } from "react";
 import Modal from "../Modal";
 
@@ -11,13 +13,13 @@ const Offers = () => {
   };
 
   return (
-    <section className="flex flex-col gap-10 bg-blue-2 p-5 pt-10">
+    <section className="bg-blue-2 flex flex-col gap-10 p-5 pt-10">
       <div className="flex flex-col items-center gap-5 lg:flex-row lg:justify-around">
-        <div className="flex flex-col justify-center gap-5 text-white-1">
-          <h2 className="relative text-center font-h text-2xl font-bold text-white xl:text-3xl 2xl:text-4xl">
+        <div className="text-white-1 flex flex-col justify-center gap-5">
+          <h2 className="font-h relative text-center text-2xl font-bold text-white xl:text-3xl 2xl:text-4xl">
             Мои предложения:
           </h2>
-          <ol className="ml-5 list-decimal font-bold text-gray-4">
+          <ol className="text-gray-4 ml-5 list-decimal font-bold">
             <li>
               <span
                 className="cursor-pointer hover:text-white"
@@ -30,7 +32,7 @@ const Offers = () => {
                 {`Индивидуальные `}
               </span>
               <span className="font-normal text-white">
-                занятия продолжительностью от 1,5 до 4 часов.
+                занятия продолжительностью от 2 до 4 часов.
               </span>
             </li>
             <li>
@@ -88,28 +90,16 @@ const Offers = () => {
                 VIP День
               </span>
             </li>
-            <li>
-              <span
-                className="cursor-pointer hover:text-white"
-                onClick={() => {
-                  document.getElementById("video")?.scrollIntoView({
-                    behavior: "smooth",
-                  });
-                }}
-              >
-                Видеосъёмка
-              </span>
-            </li>
           </ol>
           <div>
             <p>В программу индивидуальных и групповых занятий могут входить:</p>
-            <ul className="list-inside list-disc font-bold text-gray-4">
+            <ul className="text-gray-4 list-inside list-disc font-bold">
               <li
                 className="cursor-pointer hover:text-white"
                 onClick={() =>
                   clickModalHandler(
                     <>
-                      <h3 className="mb-3 text-center font-h text-2xl font-bold text-blue-4">
+                      <h3 className="font-h text-blue-4 mb-3 text-center text-2xl font-bold">
                         Базовая техника
                       </h3>
                       <p>
@@ -122,7 +112,7 @@ const Offers = () => {
                         удовольствием кататься на трассах начального уровня
                         (зелёные, не крутые синие).
                       </p>
-                    </>
+                    </>,
                   )
                 }
               >
@@ -134,7 +124,7 @@ const Offers = () => {
                 onClick={() =>
                   clickModalHandler(
                     <>
-                      <h3 className="mb-3 text-center font-h text-2xl font-bold text-blue-4">
+                      <h3 className="font-h text-blue-4 mb-3 text-center text-2xl font-bold">
                         Продвинутая техника
                       </h3>
                       <p>
@@ -145,7 +135,7 @@ const Offers = () => {
                         движения, технично и безопасно по трассам среднего
                         уровня (синие, красные).
                       </p>
-                    </>
+                    </>,
                   )
                 }
               >
@@ -157,7 +147,7 @@ const Offers = () => {
                 onClick={() =>
                   clickModalHandler(
                     <>
-                      <h3 className="mb-3 text-center font-h text-2xl font-bold text-blue-4">
+                      <h3 className="font-h text-blue-4 mb-3 text-center text-2xl font-bold">
                         Карвинг
                       </h3>
                       <p>
@@ -170,7 +160,7 @@ const Offers = () => {
                         скоростью. Постановка карвинга, при правильной
                         постановке базовой техники, требует только времени.
                       </p>
-                    </>
+                    </>,
                   )
                 }
               >
@@ -182,7 +172,7 @@ const Offers = () => {
                 onClick={() =>
                   clickModalHandler(
                     <>
-                      <h3 className="mb-3 text-center font-h text-2xl font-bold text-blue-4">
+                      <h3 className="font-h text-blue-4 mb-3 text-center text-2xl font-bold">
                         Катание по неподготовленному склону
                       </h3>
                       <p>
@@ -195,7 +185,7 @@ const Offers = () => {
                         другими лыжниками бугры, лед, &quot;расколбас&quot; и
                         пр.).
                       </p>
-                    </>
+                    </>,
                   )
                 }
               >
