@@ -1,8 +1,10 @@
+"use client";
+
 import { type ReactNode, useState, useEffect } from "react";
-import Footer from "../Footer";
+import Footer from "../../widgets/Footer";
 import Image, { type StaticImageData } from "next/image";
-import Navbar from "./Navbar";
-import Sidebar from "./Sidebar";
+import Navbar from "../../widgets/Navbar";
+import Sidebar from "../../widgets/Sidebar";
 import ScrollToTop from "react-scroll-to-top";
 import { IoIosArrowUp } from "react-icons/io";
 
@@ -25,10 +27,10 @@ const MainLayout = ({ children, imgSrc, alt, title }: MainLayoutProps) => {
 
   return (
     <>
-      <main className="relative flex min-h-screen w-full flex-col ">
+      <main className="relative flex min-h-screen w-full flex-col">
         <Navbar sidebarOpen={sidebarOpen} setSidebarOpen={setSidebarOpen} />
         <Sidebar sidebarOpen={sidebarOpen} setSidebarOpen={setSidebarOpen} />
-        <div className="absolute z-10 h-20 w-full bg-gradient-to-b from-[#2A3061] to-transparent opacity-50 sm:h-52" />
+        <div className="absolute z-10 h-20 w-full bg-linear-to-b from-[#2A3061] to-transparent opacity-50 sm:h-52" />
         <div className="absolute h-[70vh] w-full select-none lg:h-[90vh] 2xl:h-[90vh]">
           <Image
             src={imgSrc}
@@ -43,12 +45,12 @@ const MainLayout = ({ children, imgSrc, alt, title }: MainLayoutProps) => {
 
         <div className="absolute z-10 mt-[55vh] flex w-full flex-col lg:mt-[70vh] xl:mt-[80vh] 2xl:mt-[80vh]">
           <div
-            className="relative flex h-[35vh] items-center justify-center bg-blue-3 px-5"
+            className="bg-blue-3 relative flex h-[35vh] items-center justify-center px-5"
             style={{
               clipPath: "polygon(0 0, 0 86%, 100% 100%, 100% 29%)",
             }}
           >
-            <h1 className="relative z-[1] text-center font-h text-3xl font-bold text-white xl:text-4xl 2xl:text-5xl">
+            <h1 className="font-h relative z-1 text-center text-3xl font-bold text-white xl:text-4xl 2xl:text-5xl">
               {title}
             </h1>
             <div
